menambahkan pengecekan kesesuaian id user terkini dan id pemilik data

+++ app/Http/Controllers/CutiController.php
@@ -67,6 +67,9 @@ class CutiController extends Controller
 
     public function edit($id) {
         $cuti = Cuti::find($id);
+        if (!(Auth::id() == $cuti->karyawan->id || Auth::user()->role == 1)) {
+            return redirect()->back()->with('error', 'Maaf, Anda bukan pemohon data cuti ini!');
+        }
         $verifikator1 = User::where('id', '=', $cuti->verifikator_1_id)->first();
         $verifikator2 = User::where('id', '=', $cuti->verifikator_2_id)->first();
         return view('cuti.edit', compact('cuti','verifikator1','verifikator2'));
@@ -75,6 +78,9 @@ class CutiController extends Controller
     public function update(Request $request, $id) {
         // dd($request);
         $cuti = Cuti::find($id);
+        if (!(Auth::id() == $cuti->karyawan->id || Auth::user()->role == 1)) {
+            return redirect()->back()->with('error', 'Maaf, Anda bukan pemohon data cuti ini!');
+        }
         $cuti->status = 1;
         if ($request->get('tanggal_mulai') == $request->get('tanggal_akhir')) {
             return redirect()->back()->with('error', 'Tanggal Tidak Boleh Sama!');
@@ -110,6 +116,9 @@ class CutiController extends Controller
 
     public function destroy($id) {
         $cuti = Cuti::find($id);
+        if (!(Auth::id() == $cuti->karyawan->id || Auth::user()->role == 1)) {
+            return redirect()->back()->with('error', 'Maaf, Anda bukan pemohon data cuti ini!');
+        }
         // dd($cuti);
         $cuti->delete();
         return redirect()->route('cuti');
@@ -143,6 +152,9 @@ class CutiController extends Controller
 
     public function invalid($id) {
         $cuti = Cuti::find($id);
+        if (!(Auth::id() == $cuti->karyawan->id || Auth::user()->role == 1)) {
+            return redirect()->route('cuti')->with('error', 'Maaf, Anda bukan pemohon data cuti ini!');
+        }
         $cuti->status = 4;
         $cuti->update();
         return redirect()->route('cuti');
